+++ src/js/components/lightbox-panel.js
@@ -37,8 +37,8 @@ export default {
                         <div class="uk-lightbox-toolbar uk-position-top uk-text-right uk-transition-slide-top uk-transition-opaque">
                             <button class="uk-lightbox-toolbar-icon uk-close-large" type="button" uk-close></button>
                          </div>
-                        <a class="uk-lightbox-button uk-position-center-left uk-position-medium uk-transition-fade" href="#" uk-slidenav-previous uk-lightbox-item="previous"></a>
-                        <a class="uk-lightbox-button uk-position-center-right uk-position-medium uk-transition-fade" href="#" uk-slidenav-next uk-lightbox-item="next"></a>
+                        <a class="uk-lightbox-button uk-position-center-left uk-position-medium uk-transition-fade" href uk-slidenav-previous uk-lightbox-item="previous"></a>
+                        <a class="uk-lightbox-button uk-position-center-right uk-position-medium uk-transition-fade" href uk-slidenav-next uk-lightbox-item="next"></a>
                         <div class="uk-lightbox-toolbar uk-lightbox-caption uk-position-bottom uk-text-center uk-transition-slide-bottom uk-transition-opaque"></div>
                     </div>`
     }),
@@ -169,7 +169,7 @@ export default {
 
                 e.preventDefault();
 
-                this.toggleNow(this.$el, true);
+                this.toggleElement(this.$el, true, false);
 
                 this.animation = Animations['scale'];
                 removeClass(e.target, this.clsActive);