+++ src/js/core/dropdown.js
@@ -1,7 +1,0 @@
-import Drop from './drop';
-
-export default {
-
-    extends: Drop
-
-};